backend/requirements.txt
fastapi
uvicorn
pydantic>=1.8.0
requests
python-dotenv
python-multipart